fix: many bugs
