+++ src/physics/integrate.rs
@@ -3,7 +3,17 @@
 // Author: Tingxu Chen <tic128@pitt.edu>
 // Description: <Systems for physics integration>
 use bevy::prelude::*;
-use self::components::{Velocity, NetForce, Momentum, Mass};
+use super::component::{Velocity, NetForce, Momentum, Mass};
+
+pub(super) fn integrate_force_system(
+    time: Res<Time>,
+    mut query: Query<(&mut Momentum, &NetForce)>,
+) {
+    let delta_seconds = time.delta_secs();
+    for (mut momentum, net_force) in query.iter_mut() {
+        momentum.0 += net_force.0 * delta_seconds;
+    }
+}
 
 pub(super) fn integrate_momentum_system(
     mut query: Query<(&mut Velocity, &Momentum, &Mass)>,
