+++ src/physics/mod.rs
@@ -9,15 +9,15 @@ mod integrate;
 mod net_force;
 
 mod config;
-mod components;
+pub mod component;
 
 use self::gravity::gravity_system;
+use self::integrate::integrate_force_system;
 use self::integrate::integrate_momentum_system;
 use self::integrate::integrate_velocity_system;
 use self::integrate::boundary;
 use self::net_force::clean_force_system;
 use self::net_force::collect_force_events_system;
-use self::net_force::integrate_force_system;
 
 pub struct PhysicsPlugin;
 impl Plugin for PhysicsPlugin {
@@ -27,7 +27,7 @@ impl Plugin for PhysicsPlugin {
             (
                 clean_force_system,
                 gravity_system,
-                clean_rope_force_system,
+                collect_force_events_system,
                 integrate_force_system,
                 integrate_momentum_system,
                 integrate_velocity_system,
