+++ src/player/config.rs
@@ -5,6 +5,8 @@
 
 use bevy::prelude::*;
 
+const SCALE: f32 = 64.0; // 64 pixels = 1 meter
+
 // Player control parameters
 pub const PLAYER_CONTROL_SPEED_LIMIT: f32 = 10.0 * SCALE; // in pixel/s
 pub const PLAYER_MOVE_FORCE: f32 = 1500.0 * SCALE; // in Newton = kg*pixel/s^2
