+++ src/player/mod.rs
@@ -1,4 +1,5 @@
 pub mod bundle;
 pub mod player_plugin;
+pub mod config;
 
 pub use player_plugin::PlayerPlugin;
