+++ src/rope/config.rs
@@ -3,6 +3,8 @@
 // Author: Tingxu Chen <tic128@pitt.edu>
 // Description: <Rope configuration>
 
+const SCALE: f32 = 64.0; // 64 pixels = 1 meter
+
 // Rope parameters
 pub(super) const ROPE_REST_LENGTH: f32 = 5.0 * SCALE;  // in pixel
 pub(super) const ROPE_MAX_EXTENSION: f32 = 50.0;  // Maximum extension beyond rest length
